+++ CHANGELOG.md
@@ -20,6 +20,7 @@ IMPROVEMENT
 - [core] Add tx.sell_amount to SellAllCoin tags
 - [api] Limit balance watchers to 10 clients
 - [config] Add config file
+- [config] Add GUI listen address to config
 - [docs] Update documentation
 - [tendermint] Update tendermint to v0.23.1. Fixed problem with growth of wal files.